SVR-62 Implement awarding of Register App achievement
- Adds new REGISTER MethodName.
- Adds new service method for awarding registration.
- Refactors award functionality to appropriate service.
